Go: Fix review comments
